improve unittests
