+++ test/unittests/test_shared_methods.py
@@ -2,6 +2,7 @@ import unittest
 import pandas as pd
 import numpy as np
 from knora.dsplib.utils import shared_methods
+from knora.dsplib.models.propertyelement import PropertyElement
 
 
 class TestSharedMethods(unittest.TestCase):
@@ -25,6 +26,18 @@ class TestSharedMethods(unittest.TestCase):
             i, expected_row = expected
             _, returned_row = returned
             self.assertListEqual(list(expected_row), list(returned_row), msg=f"Failed in row {i}")
+
+
+    def test_check_notna(self) -> None:
+        na_values = [None, pd.NA, np.nan, "", "  ", "-", ",", ".", "*", "!", " ⳰", " ῀ ", " ῾ ", " \n\t ", "N/A", "n/a",
+                     "<NA>", ["a", "b"], pd.array(["a", "b"]), np.array([0, 1])]
+        for na_value in na_values:
+            self.assertFalse(shared_methods.check_notna(na_value), msg=f"Failed na_value: {na_value}")
+
+        notna_values = [1, 0.1, True, False, "True", "False", r" \n\t ", "0", "_", "Ὅμηρος"]
+        notna_values.extend([PropertyElement(x) for x in notna_values])
+        for notna_value in notna_values:
+            self.assertTrue(shared_methods.check_notna(notna_value), msg=f"Failed notna_value: {notna_value}")
 
 
 if __name__ == '__main__':
